requirements.txt
streamlit==1.28.0
openai>=1.12.0
pinecone>=3.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
pypdf>=4.0.0
tiktoken>=0.5.1
python-dotenv>=1.0.0